requirements.txt
Django==5.0.4
djangorestframework==3.15.2
python-dotenv==1.0.0
requests==2.31.0
